requirements.txt
langchain==0.2.14
streamlit==1.37.1
python-dotenv==1.0.1
psycopg2==2.9.9
pandas==2.2.0